+++ packages/turnkey/core/examples/react/lib/miden-turnkey.ts
@@ -106,7 +106,7 @@ export async function createMidenTurnkeyClient(
     },
     autoSync: true,
   });
-  const accountId = await createAccont(
+  const accountId = await createAccount(
     client,
     type,
     storageMode,
@@ -116,7 +116,7 @@ export async function createMidenTurnkeyClient(
   return { client, accountId };
 }
 
-export async function createAccont(
+export async function createAccount(
   midenClient: MidenClient,
   type: AccountType,
   storageMode: AccountStorageMode,